pyproject.toml
[project]
name = "hetznerinv"
version = "0.3.0"
description = ""
authors = [{ name = "ant31", email = "2t.antoine@gmail.com" }]
requires-python = ">=3.12,<4"
readme = "README.md"
license = { text = "Apache-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "pydantic",
    "typer",
    "pyyaml",
    "typing-extensions",
    "pyyaml",
    "ant31box (>=0.2.3)",
    "click (<8.2.0)",
    "hcloud (>=2.5.1,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
]

[project.scripts]
hetznerinv = "hetznerinv.cli:app"

[dependency-groups]
dev = [
    "pyreadline",
    "requests",
    "pyright",    
    "ipython",
    "pylint",
    "pyre-check",
    "types-requests",
    "isort",
    "pytest",
    "coverage",
    "pytest-cov",
    "pytest-ordering",
    "ruff",
    "pylint-pydantic",
    "pytest-asyncio",
    "bump-my-version>=1.1.2,<2",
]

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
# asyncio_default_fixture_loop_scope="function"
asyncio_default_fixture_loop_scope = "session"


[tool.bumpversion]
current_version = "0.3.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = true
commit = false
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = 'src/hetznerinv/__init__.py'
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""
regex = true
